+++ .boltrc.js
@@ -67,6 +67,7 @@ nonImageFixtures.forEach(fixturePath => {
 });
 
 module.exports = {
+  esModules: true,
   wwwDir: adjustRelativePath(siteConfig.wwwDir),
   buildDir: adjustRelativePath(siteConfig.buildDir),
   iconDir: [],
@@ -83,7 +84,7 @@ module.exports = {
   },
   prod: true,
   sourceMaps: false,
-  enableCache: true,
+  enableCache: false,
   verbosity: 1,
   copy: [...itemsToCopy],
   alterTwigEnv: [