+++ docs-site/src/pages/pattern-lab/_patterns/01-visual-styles/100-rendering-performance/lazy-lit-test/index.ts
@@ -1,8 +1,8 @@
-import { polyfillLoader } from '@bolt/core/polyfills';
+import { polyfillLoader } from '@bolt/core-v3.x/polyfills';
 
 polyfillLoader.then(res => {
-  import(/* 
-    webpackMode: 'lazy', 
-    webpackChunkName: 'lazy-lit-test' 
+  import(/*
+    webpackMode: 'lazy',
+    webpackChunkName: 'lazy-lit-test'
   */ './lazy-lit-test');
 });