+++ packages/build-tools/__tests__/multi-lang/.boltrc.multi-lang.js
@@ -1,12 +1,14 @@
 // Older method for generating a Japanese language-specific version of the build
 const config = {
   lang: ['en', 'ja'],
+  esModules: true,
   env: 'drupal', // @todo: we really need to refactor this to allow running just whatever tasks you need (more easily / more intuitively)
   buildDir: './dist-lang/build',
   dataDir: './dist-lang/build/data',
   wwwDir: './dist-lang',
   components: {
     global: [
+      '@bolt/core',
       '@bolt/components-button',
       '@bolt/global',
     ],