+++ packages/components/bolt-list/src/list.js
@@ -1,12 +1,13 @@
+import { html, customElement } from '@bolt/element';
 import {
   defineContext,
   withContext,
   props,
   define,
   hasNativeShadowDomSupport,
-} from '@bolt/core/utils';
+} from '@bolt/core-v3.x/utils';
 import classNames from 'classnames/bind';
-import { withLitHtml, html } from '@bolt/core/renderers/renderer-lit-html';
+import { withLitHtml } from '@bolt/core-v3.x/renderers/renderer-lit-html';
 import Ajv from 'ajv';
 
 import themes from '@bolt/global/styles/06-themes/_themes.all.scss';
@@ -23,14 +24,13 @@ export const ListContext = defineContext({
   display: 'inline',
   spacing: 'none',
   inset: false,
+  nowrap: false,
   align: 'start',
   separator: 'none',
 });
 
-@define
-class BoltList extends withContext(withLitHtml()) {
-  static is = 'bolt-list';
-
+@customElement('bolt-list')
+class BoltList extends withContext(withLitHtml) {
   // provide context info to children that subscribe
   // (context + subscriber idea originally from https://codepen.io/trusktr/project/editor/XbEOMk)
   static get provides() {
@@ -43,6 +43,7 @@ class BoltList extends withContext(withLitHtml()) {
     spacing: props.string, // none | xsmall | small | medium | large | xlarge
     separator: props.string, // none | solid | dashed
     inset: props.boolean, // true | false
+    nowrap: props.boolean, // true | false
     align: props.string, // start | center | end
     valign: props.string, // start | center | end
   };
@@ -81,6 +82,7 @@ class BoltList extends withContext(withLitHtml()) {
       spacing,
       separator,
       inset,
+      nowrap,
       align,
       valign,
     } = this.validateProps(this.props);
@@ -88,6 +90,7 @@ class BoltList extends withContext(withLitHtml()) {
     this.contexts.get(ListContext).display = display || this.props.display;
     this.contexts.get(ListContext).spacing = spacing || this.props.spacing;
     this.contexts.get(ListContext).inset = inset || this.props.inset;
+    this.contexts.get(ListContext).nowrap = nowrap || this.props.nowrap;
     this.contexts.get(ListContext).align = align || this.props.align;
     this.contexts.get(ListContext).separator =
       separator || this.props.separator;
@@ -99,6 +102,7 @@ class BoltList extends withContext(withLitHtml()) {
       [`c-bolt-list--align-${align}`]: align,
       [`c-bolt-list--valign-${valign}`]: valign,
       [`c-bolt-list--inset`]: inset,
+      [`c-bolt-list--nowrap`]: nowrap,
     });
 
     if (this.slots.default) {