+++ packages/components/bolt-tabs/src/tabs.js
@@ -1,3 +1,4 @@
+import { html, customElement } from '@bolt/element';
 import {
   defineContext,
   withContext,
@@ -7,14 +8,13 @@ import {
   getUniqueId,
   whichTransitionEvent,
   waitForTransitionEnd,
-} from '@bolt/core/utils';
-import { withLitHtml, html } from '@bolt/core/renderers/renderer-lit-html';
-
+} from '@bolt/core-v3.x/utils';
+import { withLitHtml } from '@bolt/core-v3.x/renderers/renderer-lit-html';
 import classNames from 'classnames/bind';
 import styles from './tabs.scss';
 import schema from '../tabs.schema.yml';
 
-import '@bolt/core/utils/optimized-resize';
+import '@bolt/core-v3.x/utils/optimized-resize';
 
 // define which specific props to provide to children that subscribe
 export const TabsContext = defineContext({
@@ -26,10 +26,8 @@ export const TabsContext = defineContext({
 
 let cx = classNames.bind(styles);
 
-@define
-class BoltTabs extends withContext(withLitHtml()) {
-  static is = 'bolt-tabs';
-
+@customElement('bolt-tabs')
+class BoltTabs extends withContext(withLitHtml) {
   static props = {
     align: props.string,
     inset: props.string,