+++ packages/configs/browserslist-config/modern.js
@@ -1,0 +1,7 @@
+module.exports = [
+  'last 2 Chrome versions',
+  'last 2 Safari versions',
+  'last 2 iOS versions',
+  'last 2 Edge versions',
+  'Firefox ESR',
+];