+++ packages/components/bolt-code-snippet/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.15.2](https://github.com/bolt-design-system/bolt/compare/v2.15.1...v2.15.2) (2020-01-21)
+
+**Note:** Version bump only for package @bolt/components-code-snippet
+
+
+
+
+
+# [2.15.0](https://github.com/bolt-design-system/bolt/compare/v2.14.3...v2.15.0) (2020-01-17)
+
+
+### Bug Fixes
+
+* remove word wrap completely and use existing scrolling mixin ([d5b6c50](https://github.com/bolt-design-system/bolt/commit/d5b6c501b2cdc4352500222a0105948677528dde))
+
+
+### Features
+
+* code snippet scrolling updates from the original Card component branch ([22816b5](https://github.com/bolt-design-system/bolt/commit/22816b5001512e044c14701d53b3edf7f54f866d))
+
+
+
+
+
+# [2.14.0](https://github.com/bolt-design-system/bolt/compare/v2.13.3...v2.14.0) (2020-01-06)
+
+
+### Features
+
+* migrate over the customElements.define-related replacement + component dependency updates from https://github.com/boltdesignsystem/bolt/pull/1579 ([3b6f529](https://github.com/bolt-design-system/bolt/commit/3b6f529))
+
+
+
+
+
+# [2.13.0](https://github.com/bolt-design-system/bolt/compare/v2.12.1...v2.13.0) (2019-12-13)
+
+
+### Features
+
+* add SQL and Java support to the Code Snippet component ([c5278da](https://github.com/bolt-design-system/bolt/commit/c5278da))
+
+
+
+
+
 # [2.12.0](https://github.com/bolt-design-system/bolt/compare/v2.11.4...v2.12.0) (2019-11-26)
 
 **Note:** Version bump only for package @bolt/components-code-snippet