+++ packages/components/bolt-critical-css-vars/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.15.2](https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-critical-css-vars/compare/v2.15.1...v2.15.2) (2020-01-21)
+
+**Note:** Version bump only for package @bolt/components-critical-css-vars
+
+
+
+
+
+# [2.15.0](https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-critical-css-vars/compare/v2.14.3...v2.15.0) (2020-01-17)
+
+**Note:** Version bump only for package @bolt/components-critical-css-vars
+
+
+
+
+
+# [2.14.0](https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-critical-css-vars/compare/v2.13.3...v2.14.0) (2020-01-06)
+
+**Note:** Version bump only for package @bolt/components-critical-css-vars
+
+
+
+
+
+## [2.13.2](https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-critical-css-vars/compare/v2.13.1...v2.13.2) (2019-12-30)
+
+**Note:** Version bump only for package @bolt/components-critical-css-vars
+
+
+
+
+
+# [2.13.0](https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-critical-css-vars/compare/v2.12.1...v2.13.0) (2019-12-13)
+
+**Note:** Version bump only for package @bolt/components-critical-css-vars
+
+
+
+
+
+## [2.12.1](https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-critical-css-vars/compare/v2.12.0...v2.12.1) (2019-12-12)
+
+**Note:** Version bump only for package @bolt/components-critical-css-vars
+
+
+
+
+
 # [2.12.0](https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-critical-css-vars/compare/v2.11.4...v2.12.0) (2019-11-26)
 
 **Note:** Version bump only for package @bolt/components-critical-css-vars